requirements.txt
# Lightweight AgenticCyber Framework Dependencies

# Google Gemini AI SDK
google-genai>=0.2.0

# Computer Vision
opencv-python>=4.8.0

# Machine Learning
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Dashboard (optional - install separately)
# Flask==2.3.3
# Flask-Login==0.6.3
# Flask-SQLAlchemy==3.0.5
# python-dotenv==1.0.0

# Standard library dependencies (included in Python)
# - pathlib
# - csv
# - socket
# - subprocess
# - os
# - abc
# - typing
# - pickle
# - json
# - warnings


